migracion de proyecto a TS
