+++ LibraryFront/.gitignore
