+++ LibraryFront/src/store/contacts.ts
@@ -1,0 +1,6 @@
+import { create } from "zustand";
+
+export const useStore = create((set) => ({
+  contacts: [],
+  setContacts: (contactsToSet) => set({ contacts: contactsToSet }),
+}));
