--- front/src/store/contacts.js
@@ -1,6 +1,0 @@
-import { create } from "zustand";
-
-export const useStore = create((set) => ({
-  contacts: [],
-  setContacts: (contactsToSet) => set({ contacts: contactsToSet }),
-}));
